- Renamed mime functions - Using lock over cache files, to avoid multiple threads handling the same request, to collide in writing the cache file - Reduced amount of public functions declared in instance.h
git-svn-id: http://galileo.dmi.unict.it/svn/weborf/trunk@486 27c2e2b3-ca6d-4234-b4fe-3aff1b8b8c5a

+++ buffered_reader.c
@@ -29,7 +29,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 This funcion inits the struct allocating a buffer of the specified size.
 It will return 0 on success and 1 on fail.
 */
-short int buffer_init(buffered_read_t * buf, ssize_t size) {
+int buffer_init(buffered_read_t * buf, ssize_t size) {
     buf->buffer = malloc(sizeof(char) * size);
     buf->size = size;
     buffer_reset(buf);
